+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add extensions to the release
-- Add Static Configuration Header `ulog_static_config.h` feature to simplify configuration
+- Add Static Configuration Header `ulog_config.h` feature to simplify configuration. Use `ULOG_BUILD_CONFIG_HEADER_ENABLED=1` to enable it.
+- Add `ULOG_BUILD_CONFIG_HEADER_NAME` to use a static configuration header instead of default `ulog_config.h`.
 
 ### Changed